+++ src/common/collections/tenants/server/methods.js
@@ -6,20 +6,20 @@ import { Tenants } from '../index.js';
 
 Meteor.methods({
     async 'pwix_tenants_manager_tenants_delete_tenant'( entity ){
-        return await Tenants.server.deleteTenant( entity, this.userId );
+        return await Tenants.s.deleteTenant( entity, this.userId );
     },
 
     // update the managers of a tenant
     async 'pwix_tenants_manager_tenants_get_scopes'(){
-        return await Tenants.server.getScopes( this.userId );
+        return await Tenants.s.getScopes( this.userId );
     },
 
     async 'pwix_tenants_manager_tenants_set_managers'( item ){
-        return await Tenants.server.setManagers( item, this.userId );
+        return await Tenants.s.setManagers( item, this.userId );
     },
 
     // upsert a tenant
     async 'pwix_tenants_manager_tenants_upsert'( item ){
-        return await Tenants.server.upsert( item, this.userId );
+        return await Tenants.s.upsert( item, this.userId );
     }
 });